First attempt at refactoring to use the new fm restful api
Update to use the new FM API. Also fix some things that broke during
the H4 refactor.  Tested to the point that I successfully ran a --dryrun
for a new Ansel release, also tested the GET call to the urls.json method,
but haven't yet tested the PUT methods yet.